Bug Fix: logging response data is not possible because TaskDelegate.data is now internal in Alamofire
There is no way to access the response data in the function networkRequestDidComplete(), because Alamofire has changed the data variable of the TaskDelegate to internal, meaning only other classes in the Alamofire module can access it. Since this logger is in it's own module, it cannot access this variable this way and thus not print out the response data when the log level is set to debug.

A workaround would be to use sessionDelegate[task].responseData {...}, but this would just enqueue the closure and log messages would be out of order – first the request would be logged, followed by the answer and then somewhen later the response data. This is not very practical, so the recommended approach is to merely log calls and introspect any returned data in the respective code.

Because of that, the code to print the response was removed.